Add pause functionality to PlayableStage and its UI rendering
Introduced a pause mechanism with `pause` and `unpause` methods, along with a `PauseScreen` UI component. Updated `PlayableStage` and derived classes to handle paused state, including pausing updates and drawing the pause menu. Ensured destroyed stages skip updates and rendering to prevent unnecessary processing.

+++ src/stages/level_1_a_stage.cpp
@@ -66,17 +66,24 @@ namespace artifact
     }
     void Level1AStage::draw() const
     {
+        if (is_being_destroyed)
+            return;
         BeginMode2D(camera);
         PlayableStage::draw();
         EndMode2D();
+        PlayableStage::draw_ui();
     }
     void Level1AStage::update(const float deltaTime)
     {
+        if (is_being_destroyed)
+            return;
         PlayableStage::update(deltaTime);
         UpdateMusicStream(music);
     }
     void Level1AStage::destroy()
     {
+        if (is_being_destroyed)
+            return;
         PlayableStage::destroy();
         UnloadMusicStream(music);
     }
